requirements.txt
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
httplib2>=0.22
protobuf
tqdm==4.67.3
